+++ src/resources/quote/quote.service.ts
@@ -3,19 +3,35 @@ import dayjs from "dayjs";
 import { sort } from "fast-sort";
 
 import { getCurrentDate } from "#common/entities";
-import { NotFoundError } from "#common/errors";
+import { NotFoundError, ServerError } from "#common/errors";
 import { FilterOperators, PaginatedResult } from "#common/types";
-import { getSortList } from "#common/utilities";
+import { getSortList, pick } from "#common/utilities";
 import { mapToArray, paginate } from "#common/utilities";
 import { AccountEntity } from "#resources/account/account.entity";
 import { database } from "#server/database";
 import { QuoteEntity, stubQuote } from "./quote.entity";
-import { QuoteCreateBody, QuoteUpdateBody } from "./quote.types";
+import { QuoteCreateBody, QuoteResponse, QuoteUpdateBody } from "./quote.types";
 
 class QuoteService {
-  public getQuotes(account?: AccountEntity, options?: FilterOperators): PaginatedResult<QuoteEntity> {
+  private mapQuoteEntityToResponse = (quote: QuoteEntity): QuoteResponse => {
+    const accountsRef = database.data!.accounts;
+    const account = accountsRef.get(quote.accountId);
+    if (!account) throw new ServerError("Invalid account requested");
+
+    return {
+      ...quote,
+      account: pick(account, ["id", "name"])
+    };
+  };
+
+  public getQuotes(
+    account?: AccountEntity,
+    options?: FilterOperators
+  ): PaginatedResult<QuoteResponse> {
     const quotesRef = database.data!.quotes;
-    const quotes = mapToArray(quotesRef).filter((q) => account ? q.accountId === account.id : q.public);
+    const quotes = mapToArray(quotesRef)
+      .filter((q) => (account ? q.accountId === account.id : q.public))
+      .map((q) => this.mapQuoteEntityToResponse(q));
 
     const validSortKeys: (keyof QuoteEntity)[] = ["author", "createdAt"];
     const sortList = getSortList(options?.sort, validSortKeys, "-createdAt");
@@ -25,13 +41,13 @@ class QuoteService {
     return paginatedQuotes;
   }
 
-  public getQuoteOfTheDay(): QuoteEntity {
+  public getQuoteOfTheDay(): QuoteResponse {
     const quotesRef = database.data!.quotes;
     const publicQuotes = mapToArray(quotesRef).filter((q) => q.public);
     faker.seed(dayjs().unix() / 86400);
     const quote = faker.helpers.arrayElement(publicQuotes);
     faker.seed();
-    return quote;
+    return this.mapQuoteEntityToResponse(quote);
   }
 
   public getQuoteById(id: string): QuoteEntity | undefined {
@@ -39,16 +55,19 @@ class QuoteService {
     return quotesRef.get(id);
   }
 
-  public getQuote(account: AccountEntity, id: string): QuoteEntity {
+  public getQuote(account: AccountEntity, id: string): QuoteResponse {
     const quote = this.getQuoteById(id);
     if (!quote || (quote.accountId !== account.id && !quote.public)) {
       throw new NotFoundError();
     }
 
-    return quote;
+    return this.mapQuoteEntityToResponse(quote);
   }
 
-  public async createQuote(account: AccountEntity, body: QuoteCreateBody): Promise<QuoteEntity> {
+  public async createQuote(
+    account: AccountEntity,
+    body: QuoteCreateBody
+  ): Promise<QuoteResponse> {
     const quote = stubQuote({
       accountId: account.id,
       ...body,
@@ -57,10 +76,13 @@ class QuoteService {
     database.data?.quotes.set(quote.id, quote);
     await database.write();
 
-    return quote;
+    return this.mapQuoteEntityToResponse(quote);
   }
 
-  public async deleteQuote(account: AccountEntity, id: string): Promise<QuoteEntity> {
+  public async deleteQuote(
+    account: AccountEntity,
+    id: string
+  ): Promise<QuoteResponse> {
     const quote = this.getQuoteById(id);
     if (!quote || quote.accountId !== account.id) {
       throw new NotFoundError();
@@ -69,10 +91,14 @@ class QuoteService {
     database.data?.quotes.delete(id);
     await database.write();
 
-    return quote;
+    return this.mapQuoteEntityToResponse(quote);
   }
 
-  public async updateQuote(account: AccountEntity, id: string, body: QuoteUpdateBody): Promise<QuoteEntity> {
+  public async updateQuote(
+    account: AccountEntity,
+    id: string,
+    body: QuoteUpdateBody
+  ): Promise<QuoteResponse> {
     const quote = this.getQuoteById(id);
     if (!quote || quote.accountId !== account.id) {
       throw new NotFoundError();
@@ -87,10 +113,9 @@ class QuoteService {
     database.data?.quotes.set(quote.id, updatedQuote);
     await database.write();
 
-    return updatedQuote;
+    return this.mapQuoteEntityToResponse(updatedQuote);
   }
 }
 
 const singleton = new QuoteService();
 export { singleton as QuoteService };
-